+++ configuration-api/main/advancedAuth.js
@@ -20,6 +20,7 @@ const typeDefs = gql`
     icon: String!
     primaryColor: String!
     secondaryColor: String!
+    customImage: String
   }
   type UserPreferencies {
     userName: String!
@@ -39,6 +40,7 @@ const typeDefs = gql`
       icon: String!
       primaryColor: String!
       secondaryColor: String!
+      file: String
     ): [TenantConfiguration]
   }
 `;
@@ -77,7 +79,7 @@ const resolvers = {
     getTenantConfig: async (object, args, context, info) => {
       try {
         config.getLogger().info(logContext, 'getTenantConfig: %s', JSON.stringify(args));
-        return await [add(args)];
+        await [add(args)];
       } catch (err) {
         config.getLogger().error(logContext, err);
         throw new ApolloError({ data: { reason: err.message } });
@@ -86,7 +88,7 @@ const resolvers = {
     modifyTenantConfig: async (object, args, context, info) => {
       try {
         config.getLogger().info(logContext, 'modifyTenantConfig: %s', JSON.stringify(args));
-        return await [update(args)];
+        return [await update(args)];
       } catch (err) {
         config.getLogger().error(logContext, err);
         throw new ApolloError({ data: { reason: err.message } });