+++ src/components/tenant/iconList.js
@@ -6,6 +6,7 @@ import AddModeratorIcon from '@mui/icons-material/AddModerator';
 export default function IconList() {
   return [
     { name: 'none', icon: '' },
+    { name: 'custom', icon: '' },
     { name: 'copy', icon: <ContentCopy fontSize="small" /> },
     { name: 'accessibility', icon: <AccessibilityIcon fontSize="small" /> },
     { name: 'shield', icon: <AddModeratorIcon fontSize="small" /> }