+++ RELEASE_NOTES.md
@@ -6,8 +6,11 @@
 
 - CI for the graphql part
 - The last tenant selected would be displayed
+- Possibility do add custom icons (with some size limitaion) to your tenant
 
 ### Bug fixes
+
+- Small scaling of the lateral menu
 
 ### Documentation